refactor: Extract OutputSchemaEditor into separate component
